(不具合) レビュー編集画面　画像変更時 #5

+++ myProject/app/Http/Controllers/editMyReviewController.php
@@ -30,13 +30,18 @@ class editMyReviewController extends Controller
 
         //更新ボタン押下時
         if (Input::get('upd-btn')) {
+            //画像変更ありの場合、ストレージの画像とreviewの画像パスを更新する
             if ($_FILES['photo']['size'] > 0) {
-                Storage::delete('public/profile_images/review-' . $selectedReviewId. '.jpg');
+                if (!is_null($reviewWork->photo_path)) {
+                    Storage::delete('public/profile_images/review-' . $selectedReviewId. '.jpg');
+                }
                 Storage::putFileAs(
                     'public/profile_images',
                     $request->photo,
                     'review-'.$selectedReviewId .'.jpg'
                 );
+                //画像なしから画像ありへの変更では、ここではじめてDBにphoto_pathが登録される
+                $reviewWork->photo_path=asset('storage/profile_images/review-' . $selectedReviewId. '.jpg');
             }
 
             $reviewWork->genre=$request->input('genre');
@@ -44,15 +49,14 @@ class editMyReviewController extends Controller
             $reviewWork->chysh=$request->input('chysh');
             $reviewWork->hyk=$request->input('hyk');
             $reviewWork->review_niy=$request->input('review_niy');
-            
-            
             $reviewWork->save();
+
             return view(
                 'common.success',
                 ['success_message'=>'レビューが更新されました',
                 'url'=>'/home']
             );
-        //削除ボタン押下時
+        //削除ボタン押下時、ストレージの画像とreviewを削除する
         } elseif (Input::get('del-btn')) {
             if (!is_null($reviewWork->photo_path)) {
                 Storage::delete('public/profile_images/review-' . $selectedReviewId. '.jpg');
